Changed the protocol to work with only 8 bits, and the data gets sent in the correct order.
This should improve the speed at which the data is sent, so we can transmit the data at a lower speed if necessary.  You're not going to notice 4096 levels of intensity.  We can still do brightness correction on the slave if necessary.

Previously I was sending the data backwards, which explained why the last 4 lights weren't coming on (I wired up pins 1-12, but I was displaying 14-5).